add get signal by id  and rename applicaiton to SignalHandlerImpl as Component class
